FIX: reordenamiento de test, documentación y bugfix login
