+++ src/main.py
@@ -22,7 +22,22 @@ from src.repositories.cards_repository import *
 from src.routers import game, home
 
 Base.metadata.create_all(bind=engine)
-app = FastAPI()
+app = FastAPI(
+        openapi_tags=[
+        {
+            "name": "base",
+            "description": "Operaciones inciales o auxiliares para el sistema.",
+        },
+        {
+            "name": "home",
+            "description": "Operaciones relacionadas con la gestión de lobbies.",
+        },
+        {
+            "name": "game",
+            "description": "Gestión de partidas, desde la creación hasta la finalización.",
+        },
+    ]
+)
 
 # Configuración de CORS
 app.add_middleware(
@@ -45,7 +60,7 @@ async def validation_exception_handler(request: Request, exc: ValidationError):
     """
     return HTTPException(status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, detail=exc.errors())
 
-@app.get("/")
+@app.get("/", tags=["base"])
 def read_root():
     """
     Verifica que el back este andando
@@ -53,7 +68,7 @@ def read_root():
     return {"mensaje": "¡Hola, FastAPI!"}
 
 
-@app.post("/login", response_model=PlayerId, status_code=status.HTTP_201_CREATED)
+@app.post("/login", tags=["base"], response_model=PlayerId, status_code=status.HTTP_201_CREATED)
 async def login(user: User, db: Session = Depends(get_db)):
     """
     Descripción: maneja la logica de logearse como usuario. Crea una instancia
@@ -64,7 +79,7 @@ async def login(user: User, db: Session = Depends(get_db)):
     - 500: Ocurre un error interno.
     """
     try:
-        jugador = add_player(user.username, False, db)#Jugador(nombre= user.username, es_anfitrion=False)
+        jugador = add_player(user.username, False, db)
     except Exception:
         db.rollback()  # Revertir cambios en caso de error
         raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Error al crear el usuario.")
