+++ src/models/cartafigura.py
@@ -52,8 +52,6 @@ class PictureCard(Base):
 
     id = Column(Integer, primary_key=True, autoincrement=True)
     figura = Column(Enum(Picture))
-    #TO DO: eliminar is_simple
-    is_simple = Column(Boolean, default=True)
     estado = Column(Enum(CardState), default=CardState.mazo)
 
     partida_id = Column(Integer, ForeignKey("partidas.id"))
